third_party/WebKit/LayoutTests/shadow-dom/css-focus-pseudo-match-shadow-host3-expected.txt

testing shadow composed tree
(1/4) both shadow hosts' delegateFocus are false
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
(2/4) top-level shadow host's delegateFocus is true
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
(3/4) lower-level shadow host's delegateFocus is true
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
(4/4) both shadow hosts' delegateFocus are true
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
